move tooltip function
