+++ superset/assets/src/visualizations/nvd3/utils.js
@@ -39,6 +39,31 @@ export const addTotalBarValues = function (svg, chart, data, stacked, axisFormat
     });
 };
 
+export function customizeToolTip(chart, xAxisFormatter, yAxisFormatters) {
+  chart.useInteractiveGuideline(true);
+  chart.interactiveLayer.tooltip.contentGenerator(function (d) {
+    const tooltipTitle = xAxisFormatter(d.value);
+    let tooltip = '';
+
+    tooltip += "<table><thead><tr><td colspan='3'>"
+      + `<strong class='x-value'>${tooltipTitle}</strong>`
+      + '</td></tr></thead><tbody>';
+
+    d.series.forEach((series, i) => {
+      const yAxisFormatter = yAxisFormatters[i];
+      const value = yAxisFormatter(series.value);
+      tooltip += "<tr><td class='legend-color-guide'>"
+        + `<div style="background-color: ${series.color};"></div></td>`
+        + `<td class='key'>${series.key}</td>`
+        + `<td class='value'>${value}</td></tr>`;
+    });
+
+    tooltip += '</tbody></table>';
+
+    return tooltip;
+  });
+}
+
 export function hideTooltips() {
   const target = document.querySelector('.nvtooltip');
   if (target) {
